requirements.txt
tqdm
pillow
pathlib
python-box
pyyaml
mlflow
dvc
dvc-s3
dagshub
pytest

 --extra-index-url https://download.pytorch.org/whl/cu118
torch
torchvision



